Fixing Small Bugs
There were a few issues here. First, open_JPK was using a strange directory to unzip all the files instead of somewhere consistent---I replaced references to the location with a variable, which is easier to change, and set it to the matlabroot. Then, the filter window was too large (legacy code error) and the filter type (unimportant for z-transform!) was set to something weird that caused an error on some data. Set it to 'none' to avoid future issues.