Pushing January Updates
A couple bug fixes, and adding MORE info to the output structure for z-transform results. It's helpful to have the z-transform indentation and force data, since they are what was used to actually do the analysis. This avoids needing to make the exact same assumptions later to re-do the transform if you want to show the raw z-domain data---and avoid issues if you can't remember what settings you used for the analysis!