Trying to save the viscoZ class
For some reason, the viscoZ class isn't being saved in some situations within the processMapZ_func function. Added a line which will save it outside the function, but it didn't fix things. It appears to be an issue with compiling the code on the cluster. Opted to rewrite SLURM script to use --module=matlab and call the matlab commands without compiling beforehand. This fixed the issue entirely. Also made changes such that we can render an animation on the cluster, which because it is Linux does NOT support mpeg-4. Using Uncompressed AVI instead. We can't use getframe() either because the cluster is headless, so we worked around by using print() directly to an RGB variable and manually creating frames.