Major Clustering Advancements
Pushing updates to a variety of sections in the code base, including test scripts that can tell us which clustering method works best for individual maps. createTestMap will take simulated datasets and use them to create fake "maps" in a bullseye shape to test how well the methodology works for known cases. In addition,  I've cleaned up some of the code for fixMapTilt and processMapZ_func so they are working properly. ViscoFitZ had an issue where it would not accept height data if we were not using the thin sample correction...which is bad---we always want to accept the data if it's given. Lastly, testClustering will sweep through all the cluster variables and now singleMapClustering will save the resulting pixel groups into the results structure for later analysis. It will overwrite previous results ONLY for the observable used in the clustering if it already exists in the structure. The date and time are also given so we know when each clustering result was stored.