Fixed reversed try-catch and for organization
Moved the try-catch for a bunch of functions beneath the for-loop that went through the subfolders for analysis. That way, we actually skip to the next directory and try if there's an error in one of them instead of killing the analysis altogether (i.e. missing all the remaining folders after the problematic one).