Found how to get the Medoids!
Returning the resulting "optimal" medoids from the analysis. This lets us actually VISUALIZE what the optimal "medoid" of each bin is. This will be an actual curve from the dataset, unlike kmeans where the centroid would have been the literal mean of all results (and thus not necessarily be an actual curve in the dataset).